test: move the local-LLM cases beside the view each one exercises (#5720)
LocalLlmTab.test.jsx was 1193 lines covering both views through the tab.
Each case now lives beside the component it drives — runtime servers,
llama-server management and the hardware profiles in
LocalLlmRuntimesView.test.jsx; the catalog, installed models, fit badges and
recommendations in LocalLlmLibraryView.test.jsx — with the assertions
unchanged, so they still pin the same behaviour.

LocalLlmTab.test.jsx keeps only what the dispatcher owns: which view a given
`view` prop mounts (including the fallback to runtimes for an absent or
unknown value, which is easy to lose in a refactor) and the shareable URLs
the pills navigate to.

Two new cases: each view asserts it leaves exactly one subscriber on the
shared `localLlm:progress` channel, and the Model Library asserts the Ollama
auto-upgrade banner now appears beside the install that triggers it.
